rename package: standards -> chapters
