+++ app/src/main/java/io/c0tt0n/realestateappraiser/ui/chapters/di/StandardsModule.kt
@@ -1,10 +1,10 @@
-package io.c0tt0n.realestateappraiser.ui.standards.di
+package io.c0tt0n.realestateappraiser.ui.chapters.di
 
 import dagger.Module
 import dagger.Provides
 import io.c0tt0n.realestateappraiser.di.scope.ActivityScope
-import io.c0tt0n.realestateappraiser.ui.standards.StandardsContract
-import io.c0tt0n.realestateappraiser.ui.standards.StandardsPresenter
+import io.c0tt0n.realestateappraiser.ui.chapters.StandardsContract
+import io.c0tt0n.realestateappraiser.ui.chapters.StandardsPresenter
 
 @Module
 class StandardsModule {
